backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Validation & Serialization
pydantic==2.10.4
pydantic[email]==2.10.4

# Utils
python-dotenv==1.0.1
python-dateutil==2.9.0.post0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1

# Development
black==24.10.0
flake8==7.1.1
pre-commit==4.0.1